Sprint 18 - исправлены замечвния и учтены рекомендации ревью
