+++ app/src/main/java/com/example/playlistmaker/albumslist/MediaLibraryActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.library.ui
+package com.example.playlistmaker.albumslist
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
