+++ app/src/main/java/com/example/playlistmaker/albumslist/MediaLibraryPagerAdapter.kt
@@ -1,11 +1,11 @@
-package com.example.playlistmaker.library.ui
+package com.example.playlistmaker.albumslist
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.example.playlistmaker.library.ui.fragments.AlbumsFragmentContener
-import com.example.playlistmaker.library.ui.fragments.FavoriteTracksFragmentContener
+import com.example.playlistmaker.albumslist.ui.fragments.AlbumsFragment
+import com.example.playlistmaker.favoritetracks.ui.fragments.FavoriteTracksFragment
 
 class MediaLibraryPagerAdapter(fragmentManager: FragmentManager, lifecycle: Lifecycle) :
     FragmentStateAdapter(fragmentManager, lifecycle) {
@@ -13,8 +13,8 @@ class MediaLibraryPagerAdapter(fragmentManager: FragmentManager, lifecycle: Life
 
     override fun createFragment(position: Int): Fragment {
        return when(position){
-            0 -> FavoriteTracksFragmentContener.newInstance()
-            else -> AlbumsFragmentContener.newInstance()
+            0 -> FavoriteTracksFragment.newInstance()
+            else -> AlbumsFragment.newInstance()
        }
     }
