+++ app/src/main/java/com/example/playlistmaker/albumslist/di/AlbumsViewModelModule.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.albumslist.di
+
+import com.example.playlistmaker.albumslist.ui.viewmodel.AlbumsListViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val albumsViewModelModule = module {
+
+    viewModel<AlbumsListViewModel> { AlbumsListViewModel() }
+
+}
