+++ app/src/main/java/com/example/playlistmaker/app/App.kt
@@ -2,7 +2,8 @@ package com.example.playlistmaker.app
 
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
-import com.example.playlistmaker.library.di.libraryViewModelModule
+import com.example.playlistmaker.albumslist.di.albumsViewModelModule
+import com.example.playlistmaker.favoritetracks.di.favoriteTracksViewModelViewModelModule
 import com.example.playlistmaker.player.di.playerDomainModule
 import com.example.playlistmaker.player.di.playerViewModelModule
 import com.example.playlistmaker.search.di.searchDataModule
@@ -41,7 +42,9 @@ class App : Application() {
                 sharingDomainModule,
                 settingsViewModelModule,
 
-                libraryViewModelModule
+                albumsViewModelModule,
+
+                favoriteTracksViewModelViewModelModule
             )
         }
         val switchPreference = getSharedPreferences(APP_SETTINGS_PREF_KEY, MODE_PRIVATE)
