+++ app/src/main/java/com/example/playlistmaker/favoritetracks/di/FavoriteTracksViewModelModule.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.favoritetracks.di
+
+import com.example.playlistmaker.favoritetracks.ui.viewmodel.FavoriteTracksViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val favoriteTracksViewModelViewModelModule = module {
+
+    viewModel<FavoriteTracksViewModel> { FavoriteTracksViewModel() }
+
+}
