+++ app/src/main/res/drawable/etit_text_cursor.xml
@@ -1,0 +1,6 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle" >
+    <size android:width="2dp"
+        android:height="20dp"/>
+    <solid android:color="#3772E7"/>
+</shape>
