+++ app/src/main/res/layout/activity_find.xml
@@ -32,6 +32,7 @@
             android:background="@drawable/rounded_edittext"
             android:drawableStart="@drawable/search"
             android:drawablePadding="9dp"
+            android:textCursorDrawable="@drawable/etit_text_cursor"
             android:drawableTint="?attr/colorSecondary"
             android:fontFamily="@font/ys_display_regular"
             android:hint="@string/bot_find"
